Refactor management of encoding errors in database tasks
Comme les InputBufferedFiled gerent desormais le nombre d'erreur d'encodage, la gestion des erreurs d'encodage
dans les  KWDatabase et KWDatabaseTask est unifiee et simplifies

Impacts principaux
- KWDataTableDriver
  - SetEncodingErrorNumber: methode supprimee, ainsi que la variable nEncodingErrorNumber
  - GetEncodingErrorNumber: methode virtruelle a redefinir
- KWDataTableDriverTextFile
  - GetEncodingErrorNumber: methode redfinie, par redirection sur le inputBuffer
  - Read, SkipMainRecord: ne gere plus les erreurs d'encodsage, puisque c'est sous-traite au inputBuffer
- PLDataTableDriverTextFile
  - GetEncodingErrorNumber: redefinition pour etre appelable meme en l'absence de buffer
- KWDatabaseTask
  - SlaveProcessStartDatabase: plus besoin d'initialiser le EncodingErrorNumber
  - SlaveProcessExploitDatabase: collecte des erreur d'encodage en debut et fin, pour retourner par difference
    les erreurs detectectees lors du SlaveProcess
  - RunDatabaseTask: memorisation des erreurs d'encodage, y compris celles specifiques aux tables externes
- KWDatabase
  - GetEncodingErrorNumber: methode desormais virtual, redefinie en lecture dans les sous-classe
  - Close: memorisation des erreurs d'encodage pour les avoir disponible apres la fermeture de la base
- KWSTDatabase
  - GetEncodingErrorNumber: methode redefinie en interrogeant le  driver
- KWMTDatabase
  - GetEncodingErrorNumber: methode redefinie en interrogeant les  drivers
  - Get|SetExternalTablesEncodingErrorNumber: methodes dedies aux tables externes
  - DMTMPhysicalComputeEncodingErrorNumber: interogation des drivers d'un mapping et de sa composition

+++ src/Norm/base/InputBufferedFile.cpp
@@ -645,7 +645,7 @@ boolean InputBufferedFile::GetNextField(char*& sField, int& nFieldLength, int& n
 		i -= iStart;
 
 		// Copie de la fin du champ (y compris le '\0')
-		memmove(sField, &sField[iStart], i + 1);
+		memmove(sField, &sField[iStart], (size_t)i + 1);
 	}
 	assert(sField[i] == '\0');
